+++ openapi-cli/src/test/java/io/ballerina/openapi/cmd/OpenAPICmdTest.java
@@ -583,7 +583,8 @@ public class OpenAPICmdTest extends OpenAPICommandTest {
     @Test(description = "Test for service generation with yaml contract without operationID")
     public void testForYamlContractWithoutOperationID() throws IOException {
         Path yamlContract = resourceDir.resolve(Paths.get("without_operationID.yaml"));
-        String[] args = {"--input", yamlContract.toString(), "-o", this.tmpDir.toString(), "--mode", "io/ballerina/openapi/corenew/service"};
+        String[] args = {"--input", yamlContract.toString(), "-o", this.tmpDir.toString(), "--mode",
+                "io/ballerina/openapi/corenew/service"};
         OpenApiCmd cmd = new OpenApiCmd(printStream, tmpDir, false);
         new CommandLine(cmd).parseArgs(args);
         cmd.execute();
