+++ openapi-core/src/main/java/io/ballerina/openapi/core/service/parameter/QueryParameterGenerator.java
@@ -1,0 +1,15 @@
+package io.ballerina.openapi.core.service.parameter;
+
+import io.ballerina.compiler.syntax.tree.ParameterNode;
+import io.swagger.v3.oas.models.Operation;
+import io.swagger.v3.oas.models.PathItem;
+
+import java.util.Map;
+
+public class QueryParameterGenerator extends ParameterGenerator {
+
+    @Override
+    ParameterNode generateParameterNode(Map.Entry<PathItem.HttpMethod, Operation> operation) {
+        return null;
+    }
+}
