+++ openapi-core/src/main/java/io/ballerina/openapi/core/service/response/LowResourceReturnTypeGenerator.java
@@ -1,0 +1,18 @@
+package io.ballerina.openapi.core.service.response;
+
+import io.ballerina.compiler.syntax.tree.AnnotationNode;
+import io.ballerina.compiler.syntax.tree.NodeList;
+import io.ballerina.compiler.syntax.tree.ReturnTypeDescriptorNode;
+import io.ballerina.openapi.core.generators.type.exception.OASTypeGenException;
+import io.swagger.v3.oas.models.Operation;
+import io.swagger.v3.oas.models.PathItem;
+
+import java.util.Map;
+
+public class LowResourceReturnTypeGenerator extends ReturnTypeGenerator {
+
+    @Override
+    public ReturnTypeDescriptorNode getReturnTypeDescriptorNode(Map.Entry<PathItem.HttpMethod, Operation> operation, NodeList<AnnotationNode> annotations, String path) throws OASTypeGenException {
+        return null;
+    }
+}
