+++ openapi-core/src/main/java/io/ballerina/openapi/core/service/response/ReturnTypeGenerator.java
@@ -1,0 +1,20 @@
+package io.ballerina.openapi.core.service.response;
+
+import io.ballerina.compiler.syntax.tree.AnnotationNode;
+import io.ballerina.compiler.syntax.tree.NodeList;
+import io.ballerina.compiler.syntax.tree.ReturnTypeDescriptorNode;
+import io.ballerina.openapi.core.generators.type.exception.OASTypeGenException;
+import io.swagger.v3.oas.models.OpenAPI;
+import io.swagger.v3.oas.models.Operation;
+import io.swagger.v3.oas.models.PathItem;
+
+import java.util.Map;
+
+public abstract class ReturnTypeGenerator {
+    String pathRecord;
+    OpenAPI openAPI;
+
+    public abstract ReturnTypeDescriptorNode getReturnTypeDescriptorNode(Map.Entry<PathItem.HttpMethod, Operation> operation,
+                                                         NodeList<AnnotationNode> annotations, String path)
+            throws OASTypeGenException;
+}
