+++ openapi-core/src/main/java/io/ballerina/openapi/core/service/response/ReturnTypeGeneratorImpl.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package io.ballerina.openapi.core.service;
+package io.ballerina.openapi.core.service.response;
 
 import io.ballerina.compiler.syntax.tree.AnnotationNode;
 import io.ballerina.compiler.syntax.tree.BuiltinSimpleNameReferenceNode;
@@ -32,12 +32,14 @@ import io.ballerina.compiler.syntax.tree.TypeDescriptorNode;
 import io.ballerina.openapi.core.generators.type.GeneratorUtils;
 import io.ballerina.openapi.core.generators.common.TypeHandler;
 import io.ballerina.openapi.core.generators.type.exception.OASTypeGenException;
+import io.ballerina.openapi.core.service.GeneratorConstants;
+import io.ballerina.openapi.core.service.diagnostic.ServiceDiagnosticMessages;
+import io.ballerina.openapi.core.service.ServiceGenerationUtils;
 import io.swagger.v3.oas.models.OpenAPI;
 import io.swagger.v3.oas.models.Operation;
 import io.swagger.v3.oas.models.PathItem;
 import io.swagger.v3.oas.models.media.Content;
 import io.swagger.v3.oas.models.media.MediaType;
-import io.swagger.v3.oas.models.media.Schema;
 import io.swagger.v3.oas.models.responses.ApiResponse;
 import io.swagger.v3.oas.models.responses.ApiResponses;
 
@@ -66,15 +68,11 @@ import static io.ballerina.openapi.core.service.GeneratorConstants.RETURNS;
  *
  * @since 1.3.0
  */
-public class ReturnTypeGenerator {
-
-    private final String pathRecord;
-    private String httpMethod;
-    private OpenAPI openAPI;
+public class ReturnTypeGeneratorImpl extends ReturnTypeGenerator {
 
     private static int countForRecord = 0;
 
-    public ReturnTypeGenerator(String pathRecord, OpenAPI openAPI) {
+    public ReturnTypeGeneratorImpl(String pathRecord, OpenAPI openAPI) {
         this.pathRecord = pathRecord;
         this.openAPI = openAPI;
     }
@@ -92,12 +90,13 @@ public class ReturnTypeGenerator {
             throws OASTypeGenException {
         ReturnTypeDescriptorNode returnNode;
         List<String> returnDescriptions = new ArrayList<>();
-        httpMethod = operation.getKey().name().toLowerCase(Locale.ENGLISH);
+        String httpMethod = operation.getKey().name().toLowerCase(Locale.ENGLISH);
         if (operation.getValue().getResponses() != null) {
             ApiResponses responses = operation.getValue().getResponses();
             if (responses.size() > 1) {
                 //handle multiple response scenarios ex: status code 200, 400, 500
-                TypeDescriptorNode type = handleMultipleResponse(responses, returnDescriptions, httpMethod, openAPI, pathRecord);
+                TypeDescriptorNode type = handleMultipleResponse(responses, returnDescriptions, httpMethod, openAPI,
+                        pathRecord);
                 returnNode = createReturnTypeDescriptorNode(createToken(RETURNS_KEYWORD), annotations, type);
             } else if (responses.size() == 1) {
                 //handle single response
@@ -135,7 +134,8 @@ public class ReturnTypeGenerator {
     /**
      * Generate union type node when operation has multiple responses.
      */
-    public static TypeDescriptorNode handleMultipleResponse(ApiResponses responses, List<String> returnDescription, String httpMethod, OpenAPI openAPI, String pathRecord)
+    private static TypeDescriptorNode handleMultipleResponse(ApiResponses responses, List<String> returnDescription,
+                                                            String httpMethod, OpenAPI openAPI, String pathRecord)
             throws OASTypeGenException {
 
         Set<String> qualifiedNodes = new LinkedHashSet<>();
@@ -148,7 +148,8 @@ public class ReturnTypeGenerator {
             String typeName = null;
 
             if (code == null && !responseCode.equals(GeneratorConstants.DEFAULT)) {
-                throw new OASTypeGenException(String.format(ServiceDiagnosticMessages.OAS_SERVICE_107.getDescription(), responseCode));
+                throw new OASTypeGenException(String.format(ServiceDiagnosticMessages.OAS_SERVICE_107.getDescription(),
+                        responseCode));
             }
             if (responseValue != null && responseValue.get$ref() != null) {
                 String[] splits = responseValue.get$ref().split("/");
@@ -162,26 +163,30 @@ public class ReturnTypeGenerator {
             } else if (content == null && (responseValue == null || responseValue.get$ref() == null) ||
                     content != null && content.size() == 0) {
                 //key and value
-                QualifiedNameReferenceNode node = ServiceGenerationUtils.getQualifiedNameReferenceNode(GeneratorConstants.HTTP,
-                        code);
+                QualifiedNameReferenceNode node = ServiceGenerationUtils.getQualifiedNameReferenceNode(
+                        GeneratorConstants.HTTP, code);
                 typeName = node.toSourceCode();
             } else if (content != null) {
                 TypeDescriptorNode bodyType = handleMultipleContents(content.entrySet(), pathRecord);
                 //Check the default behaviour for return type according to POST method.
                 boolean isWithOutStatusCode =
-                        (httpMethod.equals(GeneratorConstants.POST) && responseCode.equals(GeneratorConstants.HTTP_201)) ||
-                                (!httpMethod.equals(GeneratorConstants.POST) && responseCode.equals(GeneratorConstants.HTTP_200));
+                        (httpMethod.equals(GeneratorConstants.POST) &&
+                                responseCode.equals(GeneratorConstants.HTTP_201)) ||
+                                (!httpMethod.equals(GeneratorConstants.POST) &&
+                                        responseCode.equals(GeneratorConstants.HTTP_200));
                 if (isWithOutStatusCode) {
                     typeName = bodyType.toSourceCode();
                 } else {
-                    SimpleNameReferenceNode node = TypeHandler.getInstance().createStatusCodeTypeInclusionRecord(code, bodyType);
+                    SimpleNameReferenceNode node = TypeHandler.getInstance()
+                            .createStatusCodeTypeInclusionRecord(code, bodyType);
                     typeName = node.name().text();
                 }
             }
             if (typeName != null) {
                 qualifiedNodes.add(typeName);
                 if (responseValue.getDescription() != null && !responseValue.getDescription().isEmpty()) {
-                    returnDescription.add(typeName.trim() + GeneratorConstants.PIPE + responseValue.getDescription().trim());
+                    returnDescription.add(typeName.trim() + GeneratorConstants.PIPE +
+                            responseValue.getDescription().trim());
                 }
             }
         }
@@ -196,19 +201,20 @@ public class ReturnTypeGenerator {
     /**
      * Generate union type node when response has multiple content types.
      */
-    public static TypeDescriptorNode handleMultipleContents(Set<Map.Entry<String, MediaType>> contentEntries, String pathRecord)
+    private static TypeDescriptorNode handleMultipleContents(Set<Map.Entry<String, MediaType>> contentEntries,
+                                                            String pathRecord)
             throws OASTypeGenException {
         Set<String> qualifiedNodes = new LinkedHashSet<>();
         for (Map.Entry<String, MediaType> contentType : contentEntries) {
             String recordName = getNewRecordName(pathRecord);
-            TypeDescriptorNode mediaTypeToken = generateTypeDescriptorForMediaTypes(contentType, recordName);
+            TypeDescriptorNode mediaTypeToken = ServiceGenerationUtils
+                    .generateTypeDescriptorForMediaTypes(contentType, recordName);
             if (mediaTypeToken == null) {
-                SimpleNameReferenceNode httpResponse = createSimpleNameReferenceNode(createIdentifierToken(GeneratorConstants.ANYDATA));
+                SimpleNameReferenceNode httpResponse = createSimpleNameReferenceNode(createIdentifierToken(
+                        GeneratorConstants.ANYDATA));
                 qualifiedNodes.add(httpResponse.name().text());
-            } else if (mediaTypeToken instanceof NameReferenceNode nameReferenceNode) {
-//                BallerinaTypesGenerator.typeDefinitionNodes.put(rightNode.get().typeName().text(), rightNode.get());
+            } else if (mediaTypeToken instanceof NameReferenceNode) {
                 setCountForRecord(countForRecord++);
-//                qualifiedNodes.add(TypeHandler.getSimpleNameReferenceNode(recordName).toSourceCode());
                 qualifiedNodes.add(mediaTypeToken.toSourceCode());
             } else {
                 qualifiedNodes.add(mediaTypeToken.toSourceCode());
@@ -228,8 +234,9 @@ public class ReturnTypeGenerator {
     /**
      * This util is to generate return node when the operation has one response.
      */
-    public ReturnTypeDescriptorNode handleSingleResponse(NodeList<AnnotationNode> annotations,
-                                                         Map.Entry<String, ApiResponse> response, String pathRecord, String httpMethod)
+    private ReturnTypeDescriptorNode handleSingleResponse(NodeList<AnnotationNode> annotations,
+                                                         Map.Entry<String, ApiResponse> response, String pathRecord,
+                                                          String httpMethod)
             throws OASTypeGenException {
 
         ReturnTypeDescriptorNode returnNode = null;
@@ -253,7 +260,8 @@ public class ReturnTypeGenerator {
             String responseCode = response.getKey().trim();
             boolean isWithOutStatusCode =
                     (httpMethod.equals(GeneratorConstants.POST) && responseCode.equals(GeneratorConstants.HTTP_201)) ||
-                            (!httpMethod.equals(GeneratorConstants.POST) && responseCode.equals(GeneratorConstants.HTTP_200));
+                            (!httpMethod.equals(GeneratorConstants.POST) &&
+                                    responseCode.equals(GeneratorConstants.HTTP_200));
             if (isWithOutStatusCode) {
                 // handle 200, 201 status code
                 Set<Map.Entry<String, MediaType>> contentEntries = responseContent.entrySet();
@@ -261,7 +269,8 @@ public class ReturnTypeGenerator {
                 if (contentEntries.size() > 1) {
                     returnType = handleMultipleContents(contentEntries, pathRecord);
                 } else {
-                    returnType = getReturnNodeForSchemaType(contentEntries.iterator().next(), getNewRecordName(pathRecord));
+                    returnType = getReturnNodeForSchemaType(contentEntries.iterator().next(),
+                            getNewRecordName(pathRecord));
                 }
                 returnNode = createReturnTypeDescriptorNode(returnKeyWord, annotations, returnType);
             } else if (response.getKey().trim().equals(GeneratorConstants.DEFAULT)) {
@@ -282,7 +291,8 @@ public class ReturnTypeGenerator {
                     Iterator<Map.Entry<String, MediaType>> contentItr = responseContent.entrySet().iterator();
                     Map.Entry<String, MediaType> mediaTypeEntry = contentItr.next();
                     String recordName = getNewRecordName(pathRecord);
-                    TypeDescriptorNode mediaTypeToken = generateTypeDescriptorForMediaTypes(mediaTypeEntry, recordName);
+                    TypeDescriptorNode mediaTypeToken = ServiceGenerationUtils
+                            .generateTypeDescriptorForMediaTypes(mediaTypeEntry, recordName);
 
                     if (mediaTypeToken == null) {
                         type = createSimpleNameReferenceNode(createIdentifierToken(GeneratorConstants.ANYDATA));
@@ -291,7 +301,8 @@ public class ReturnTypeGenerator {
                     }
                 }
                 if (!type.toString().equals(HTTP_RESPONSE)) {
-                    SimpleNameReferenceNode recordType = TypeHandler.getInstance().createStatusCodeTypeInclusionRecord(code, type);
+                    SimpleNameReferenceNode recordType = TypeHandler.getInstance()
+                            .createStatusCodeTypeInclusionRecord(code, type);
                     NodeList<AnnotationNode> annotation = createEmptyNodeList();
                     returnNode = createReturnTypeDescriptorNode(returnKeyWord, annotation, recordType);
                 }
@@ -300,68 +311,17 @@ public class ReturnTypeGenerator {
         return returnNode;
     }
 
-    /**
-     * Generate TypeDescriptor for all the mediaTypes.
-     *
-     * Here return the @code{ImmutablePair<Optional<TypeDescriptorNode>, TypeDefinitionNode>} for return type.
-     * Left node(key) of the return tuple represents the return type node for given mediaType details, Right Node
-     * (Value) of the return tuple represents the newly generated TypeDefinitionNode for return type if it has inline
-     * objects.
-     */
-    public static TypeDescriptorNode generateTypeDescriptorForMediaTypes(
-            Map.Entry<String, MediaType> mediaType, String recordName) throws OASTypeGenException {
-        String mediaTypeContent = selectMediaType(mediaType.getKey().trim());
-        Schema<?> schema = mediaType.getValue().getSchema();
-        switch (mediaTypeContent) {
-            case GeneratorConstants.APPLICATION_JSON:
-                TypeDescriptorNode typeDescriptorNode = TypeHandler.getInstance().generateTypeDescriptorForJsonContent(schema, recordName);
-                return typeDescriptorNode;
-            case GeneratorConstants.APPLICATION_XML:
-                typeDescriptorNode = TypeHandler.getInstance().generateTypeDescriptorForXMLContent();
-                return typeDescriptorNode;
-            case GeneratorConstants.APPLICATION_URL_ENCODE:
-                typeDescriptorNode = TypeHandler.getInstance().generateTypeDescriptorForMapStringContent();
-                return typeDescriptorNode;
-            case GeneratorConstants.TEXT:
-                typeDescriptorNode = TypeHandler.getInstance().generateTypeDescriptorForTextContent(GeneratorConstants.STRING);
-                return typeDescriptorNode;
-            case GeneratorConstants.APPLICATION_OCTET_STREAM:
-                typeDescriptorNode = TypeHandler.getInstance().generateTypeDescriptorForOctetStreamContent();
-                return typeDescriptorNode;
-            default:
-                return null;
-        }
-    }
-
-    public TypeDescriptorNode getReturnNodeForSchemaType(Map.Entry<String, MediaType> contentEntry, String recordName)
+    private TypeDescriptorNode getReturnNodeForSchemaType(Map.Entry<String, MediaType> contentEntry, String recordName)
             throws OASTypeGenException {
         TypeDescriptorNode returnNode;
-        TypeDescriptorNode mediaTypeToken = generateTypeDescriptorForMediaTypes(contentEntry, recordName);
+        TypeDescriptorNode mediaTypeToken = ServiceGenerationUtils
+                .generateTypeDescriptorForMediaTypes(contentEntry, recordName);
         if (mediaTypeToken == null) {
             returnNode = createSimpleNameReferenceNode(createIdentifierToken(GeneratorConstants.ANYDATA));
         } else {
             returnNode = mediaTypeToken;
         }
         return returnNode;
-    }
-
-    /**
-     *
-     * This util is used for selecting standard media type by looking at the user defined media type.
-     */
-    public static String selectMediaType(String mediaTypeContent) {
-        if (mediaTypeContent.matches("application/.*\\+json") || mediaTypeContent.matches(".*/json")) {
-            mediaTypeContent = GeneratorConstants.APPLICATION_JSON;
-        } else if (mediaTypeContent.matches("application/.*\\+xml") || mediaTypeContent.matches(".*/xml")) {
-            mediaTypeContent = GeneratorConstants.APPLICATION_XML;
-        } else if (mediaTypeContent.matches("text/.*")) {
-            mediaTypeContent = GeneratorConstants.TEXT;
-        }  else if (mediaTypeContent.matches("application/.*\\+octet-stream")) {
-            mediaTypeContent = GeneratorConstants.APPLICATION_OCTET_STREAM;
-        } else if (mediaTypeContent.matches("application/.*\\+x-www-form-urlencoded")) {
-            mediaTypeContent = GeneratorConstants.APPLICATION_URL_ENCODE;
-        }
-        return mediaTypeContent;
     }
 
     private static String getNewRecordName(String pathRecord) {
