+++ openapi-core/src/main/java/module-info.java
@@ -52,4 +52,7 @@ module io.ballerina.openapi.core {
     exports io.ballerina.openapi.core.generators.type.model;
     exports io.ballerina.openapi.core.generators.type.exception;
     exports io.ballerina.openapi.core.generators.common;
+    exports io.ballerina.openapi.core.service.parameter;
+    exports io.ballerina.openapi.core.service.response;
+    exports io.ballerina.openapi.core.service.diagnostic;
 }
